feat(recipes): Creates a 'button' that gives the user access to the Login page.
The user must be able to access the Login area to register or access their already created account from the home page. However, this button must only be visible on the home page and while the user is not logged in. As the button is in the header, which appears on all pages of the application, it was necessary to create a conditional rendering validating the type of page being rendered.

ref:https://github.com/deb-gama/django_recipes/issues/9