fix(recipes): Resolves recipe rendering bug that does not have a cover image
Recipes that did not have a cover were saved normally, but when published - is_published = True - they broke the rendering of the home page. In order to maintain the standardization of the application, the cover field is mandatory and the recipe registration includes the obligation to send a valid image to be published.

ref: https://github.com/deb-gama/django_recipes/issues/11